+++ CHANGELOG.md
@@ -24,6 +24,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `LocalPath::ResolveToCurrentDirectory`: effectively calculates `currentDirectory / this`. No-op for paths that are already absolute (aside from converting to the `AbsolutePath` type).
 
   Thanks to @Illusion4.
+- `AbsolutePath::ReadKind` to check the file system object kind (file, directory, or something else) and whether it exists at all.
+
+  Thanks to @Kataane.
 
 ## [1.3.0] - 2024-06-21
 ### Added